Start transition into importer framework.
