+++ deepwell/importer/constants.py
@@ -1,0 +1,3 @@
+from datetime import datetime
+
+UNKNOWN_CREATION_DATE = datetime.fromtimestamp(0)
