+++ deepwell/importer/requirements.txt
@@ -1,3 +1,2 @@
 bidict>=0.22.0
 psycopg2>=2.9.3
-py7zr>=0.19.0
